+++ tech1-framework-b2b-base-security-jwt/src/main/java/io/tech1/framework/b2b/base/security/jwt/domain/sessions/Session.java
@@ -1,10 +1,12 @@
 package io.tech1.framework.b2b.base.security.jwt.domain.sessions;
 
+import io.tech1.framework.b2b.base.security.jwt.domain.jwt.JwtAccessToken;
 import io.tech1.framework.b2b.base.security.jwt.domain.jwt.JwtRefreshToken;
 import io.tech1.framework.domain.base.Username;
 
 public record Session(
         Username username,
+        JwtAccessToken accessToken,
         JwtRefreshToken refreshToken
 ) {
 }