+++ tech1-framework-b2b-postgres-security-jwt/src/main/java/io/tech1/framework/b2b/postgres/security/jwt/domain/db/PostgresDbUserSession.java
@@ -4,18 +4,26 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.tech1.framework.b2b.base.security.jwt.domain.db.AnyDbUserSession;
 import io.tech1.framework.b2b.base.security.jwt.domain.dto.responses.ResponseUserSession2;
 import io.tech1.framework.b2b.base.security.jwt.domain.identifiers.UserSessionId;
-import io.tech1.framework.b2b.base.security.jwt.domain.jwt.CookieRefreshToken;
+import io.tech1.framework.b2b.base.security.jwt.domain.jwt.CookieAccessToken;
+import io.tech1.framework.b2b.base.security.jwt.domain.jwt.JwtAccessToken;
 import io.tech1.framework.b2b.base.security.jwt.domain.jwt.JwtRefreshToken;
+import io.tech1.framework.b2b.postgres.security.jwt.converters.columns.PostgresJwtAccessTokenConverter;
+import io.tech1.framework.b2b.postgres.security.jwt.converters.columns.PostgresJwtRefreshTokenConverter;
 import io.tech1.framework.b2b.postgres.security.jwt.converters.columns.PostgresUserRequestMetadataConverter;
 import io.tech1.framework.b2b.postgres.security.jwt.converters.columns.PostgresUsernameConverter;
 import io.tech1.framework.domain.base.Username;
 import io.tech1.framework.domain.http.requests.UserRequestMetadata;
 import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
 
+import static io.tech1.framework.b2b.base.security.jwt.domain.db.AnyDbUserSession.ofNotPersisted;
+import static io.tech1.framework.b2b.base.security.jwt.domain.db.AnyDbUserSession.ofPersisted;
 import static io.tech1.framework.b2b.postgres.security.jwt.constants.PostgreTablesConstants.USERS_SESSIONS;
+import static java.util.Objects.nonNull;
 
+@SuppressWarnings("JpaDataSourceORMInspection")
 // Lombok
 @NoArgsConstructor
 @Getter
@@ -27,59 +35,51 @@ import static io.tech1.framework.b2b.postgres.security.jwt.constants.PostgreTabl
 @Table(name = USERS_SESSIONS)
 public class PostgresDbUserSession {
     @Id
-//    @GenericGenerator(name = "uuid2", strategy = "uuid2")
-//    @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "uuid2")
-//    @Column(length = 36, nullable = false, updatable = false)
-    @Column(length = 256, nullable = false, updatable = false)
+    @GenericGenerator(name = "uuid2", strategy = "uuid2")
+    @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "uuid2")
+    @Column(length = 36, nullable = false, updatable = false)
     private String id;
 
     @Convert(converter = PostgresUsernameConverter.class)
     @Column(nullable = false)
     private Username username;
 
+    @Convert(converter = PostgresJwtAccessTokenConverter.class)
+    @Column(name = "access_token", length = 1024, nullable = false)
+    private JwtAccessToken accessToken;
+
+    @Convert(converter = PostgresJwtRefreshTokenConverter.class)
+    @Column(name = "refresh_token", length = 1024, nullable = false)
+    private JwtRefreshToken refreshToken;
+
     @Convert(converter = PostgresUserRequestMetadataConverter.class)
     @Column(length = 65535, nullable = false)
     private UserRequestMetadata metadata;
 
-    public PostgresDbUserSession(JwtRefreshToken jwtRefreshToken, Username username, UserRequestMetadata metadata) {
-        this.id = jwtRefreshToken.value();
-        this.username = username;
-        this.metadata = metadata;
-    }
-
     public PostgresDbUserSession(AnyDbUserSession session) {
-        this.id = session.id().value();
+        if (session.persisted()) {
+            this.id = session.id().value();
+        }
         this.username = session.username();
+        this.accessToken = session.accessToken();
+        this.refreshToken = session.refreshToken();
         this.metadata = session.metadata();
-    }
-
-    @Deprecated
-    @JsonIgnore
-    @Transient
-    public JwtRefreshToken getJwtRefreshToken() {
-        return new JwtRefreshToken(this.id);
     }
 
     @JsonIgnore
     @Transient
     public AnyDbUserSession anyDbUserSession() {
-        return new AnyDbUserSession(
-                new UserSessionId(this.id),
-                this.username,
-                this.metadata,
-                this.getJwtRefreshToken()
-        );
+        if (nonNull(this.id)) {
+            return ofPersisted(new UserSessionId(this.id), this.username, this.accessToken, this.refreshToken, this.metadata);
+        } else {
+            return ofNotPersisted( this.username, this.accessToken, this.refreshToken, this.metadata);
+        }
     }
 
     @JsonIgnore
     @Transient
-    public ResponseUserSession2 responseUserSession2(CookieRefreshToken cookie) {
-        return ResponseUserSession2.of(
-                this.username,
-                this.metadata,
-                this.getJwtRefreshToken(),
-                cookie
-        );
+    public ResponseUserSession2 responseUserSession2(CookieAccessToken cookie) {
+        return ResponseUserSession2.of(new UserSessionId(this.id), this.username, cookie, this.accessToken, this.metadata);
     }
 
     @JsonIgnore