+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 # Release Notes
 
 ### v1.14
-- MongoDb: truncate/delete "tech1_user_sessions"
+- MongoDb: truncate/delete "tech1_user_sessions" (MongoDbUserSession: requestMetadata → metadata)
 - Users API: /user → /users + Postman
 - Delete "sessionId" in ResponseUserSession2 
 - Add "mongodbSecurityJwtConfigs" property (migrate "securityJwtConfigs.mongodb" configs)